requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
click==8.1.8
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.3.2
propcache==0.3.1
python-dotenv==1.1.0
soupsieve==2.6
typing_extensions==4.13.1
Werkzeug==3.1.3
yarl==1.18.3
